fix: refactor roles API integration and update role fetching methods [WTEL-7204](https://webitel.atlassian.net/browse/WTEL-7204)
