+++ src/api/clients/roles/roles.ts
@@ -1,11 +1,13 @@
 import deepCopy from 'deep-copy';
+import { RolesApiFactory } from 'webitel-sdk';
 
 import ApplicationsAccess from '../../../modules/Userinfo/classes/ApplicationsAccess.js';
 import {
   getDefaultGetListResponse,
   getDefaultGetParams,
   getDefaultInstance,
-} from '../../defaults/index.js';
+  getDefaultOpenAPIConfig,
+} from '../../defaults';
 import applyTransform, {
   camelToSnake,
   generateUrl,
@@ -14,11 +16,13 @@ import applyTransform, {
   sanitize,
   snakeToCamel,
   starToSearch,
-} from '../../transformers/index.js';
+} from '../../transformers';
 
 const instance = getDefaultInstance();
+const configuration = getDefaultOpenAPIConfig();
 
-const baseUrl = '/roles';
+const rolesApiFactory = RolesApiFactory(configuration, '', instance);
+
 const fieldsToSend = ['name', 'description', 'permissions', 'metadata'];
 
 const preRequestHandler = (item) => {
@@ -28,22 +32,32 @@ const preRequestHandler = (item) => {
 };
 
 const getRoleList = async (params) => {
-  const fieldsToSend = ['page', 'size', 'q', 'sort', 'fields', 'ids'];
+  const fieldsToSend = ['page', 'size', 'q', 'sort', 'fields', 'id'];
 
-  const url = applyTransform(params, [
-    merge(getDefaultGetParams()),
-    (params) => {
-      params.ids = params.ids || params.id; // accept either ids or id as param
-      return params;
-    },
-    starToSearch('search'),
-    (params) => ({ ...params, q: params.search }),
-    sanitize(fieldsToSend),
-    camelToSnake(),
-    generateUrl(baseUrl),
-  ]);
+  const { page, size, q, sort, name, fields, id, userId, userName } =
+    applyTransform(params, [
+      merge(getDefaultGetParams()),
+      starToSearch('search'),
+      (params) => {
+        params.ids = params.ids || params.id; // accept either ids or id as param
+        return { ...params, q: params.search };
+      },
+      sanitize(fieldsToSend),
+      camelToSnake(),
+    ]);
+
   try {
-    const response = await instance.get(url);
+    const response = await rolesApiFactory.searchRoles(
+      [id],
+      name,
+      userId,
+      userName,
+      q,
+      fields,
+      sort,
+      page,
+      size,
+    );
     const { items, next } = applyTransform(response.data, [
       snakeToCamel(),
       merge(getDefaultGetListResponse()),
@@ -73,41 +87,13 @@ const getRole = async ({ itemId: id }) => {
     return copy;
   };
 
-  const url = `${baseUrl}/${id}?fields=metadata&fields=permissions&fields=name&fields=description`;
-
   try {
-    const response = await instance.get(url);
+    const response = await rolesApiFactory.readRole(id, fieldsToSend);
     return applyTransform(response.data, [
       snakeToCamel(),
       merge(defaultObject),
       itemResponseHandler,
     ]);
-  } catch (err) {
-    throw applyTransform(err, [notify]);
-  }
-};
-
-const getExtendedRoles = async (params) => {
-  const fieldsToSend = ['page', 'size', 'q', 'sort', 'fields', 'id'];
-
-  const url = applyTransform(params, [
-    merge(getDefaultGetParams()),
-    starToSearch('search'),
-    (params) => ({ ...params, q: params.search }),
-    sanitize(fieldsToSend),
-    camelToSnake(),
-    generateUrl(baseUrl),
-  ]);
-  try {
-    const response = await instance.get(url);
-    const { items, next } = applyTransform(response.data, [
-      snakeToCamel(),
-      merge(getDefaultGetListResponse()),
-    ]);
-    return {
-      items,
-      next,
-    };
   } catch (err) {
     throw applyTransform(err, [notify]);
   }
@@ -120,7 +106,7 @@ const addRole = async ({ itemInstance }) => {
     camelToSnake(),
   ]);
   try {
-    const response = await instance.post(baseUrl, item);
+    const response = await rolesApiFactory.createRole(item);
     return applyTransform(response.data, [snakeToCamel()]);
   } catch (err) {
     throw applyTransform(err, [notify]);
@@ -134,9 +120,8 @@ const updateRole = async ({ itemInstance, itemId: id }) => {
     camelToSnake(),
   ]);
 
-  const url = `${baseUrl}/${id}`;
   try {
-    const response = await instance.put(url, item);
+    const response = await rolesApiFactory.updateRole(id, item);
     return applyTransform(response.data, [snakeToCamel()]);
   } catch (err) {
     throw applyTransform(err, [notify]);
@@ -144,9 +129,8 @@ const updateRole = async ({ itemInstance, itemId: id }) => {
 };
 
 const deleteRole = async ({ id }) => {
-  const url = `${baseUrl}/${id}`;
   try {
-    const response = await instance.delete(url);
+    const response = await rolesApiFactory.deleteRole(id);
     return applyTransform(response.data, []);
   } catch (err) {
     throw applyTransform(err, [notify]);
@@ -195,7 +179,6 @@ const RolesAPI = {
   delete: deleteRole,
   getLookup: getRolesLookup,
 
-  getExtendedRoles,
   getPermissionsOptions,
 };
